requirements.txt
apache-airflow==2.4.1
scikit-learn
mlflow
psycopg2-binary
fastapi
uvicorn
